[Package] [Handlers] Cleanup Code Base
